requirements.txt
alembic==1.11.2
bcrypt==4.0.1
blinker==1.6.2
cfgv==3.3.1
click==8.1.6
distlib==0.3.7
filelock==3.12.2
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
identify==2.5.26
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
nodeenv==1.8.0
platformdirs==3.10.0
pre-commit==3.3.3
psycopg2==2.9.7
PyJWT==2.8.0
pyscopg2==66.0.2
PyYAML==6.0.1
SQLAlchemy==2.0.19
typing_extensions==4.7.1
virtualenv==20.24.2
Werkzeug==2.3.6
